requirements.txt
azure-core
semantic-kernel
semantic-kernel[azure]
semantic-kernel[mcp]
openai

# AutoGen framework
autogen-agentchat
autogen-core
autogen-ext
autogen-ext[azure]

# Azure AI services
azure-ai-inference
azure-ai-projects
azure-search-documents
httpx
ipykernel
pillow
python-dotenv
pandas
uvicorn
nest-asyncio

agent-framework


mcp[cli]

# Other dependencies
a2a-sdk
numpy